remove some unused code

+++ Utilities/ImageUtils.cs
@@ -59,9 +59,9 @@ namespace grp
         public static Image<Rgba32> GenerateImageWhichFits(IEnumerable<int> widths, IEnumerable<int> heights, MergeDirection direction, float overlap) => direction switch
         {
             MergeDirection.BottomTop or MergeDirection.TopBottom => new(widths.Max(), 
-                heights.Select(x => (int)(x * (1 - overlap))).Sum() + (int)MathUtils.Mean(heights.First(), heights.Last())),
+                heights.Select(x => (int)(x * (1 - overlap))).Sum() + MathUtils.Mean(heights.First(), heights.Last())),
             MergeDirection.LeftRight or MergeDirection.RightLeft => new(widths.Select(x => (int)(x * (1 - overlap) 
-            + (int)MathUtils.Mean(widths.First(), widths.Last()))).Sum(), heights.Max()),
+            + MathUtils.Mean(widths.First(), widths.Last()))).Sum(), heights.Max()),
             _ => throw new ArgumentOutOfRangeException(nameof(direction))
         };
         /// <summary>
